ai-career/v2/requirements.txt
pyyaml==6.0.2
numpy==1.26.4
sentence-transformers==2.7.0
torch==2.2.2